add-update-function

+++ app/Http/Controllers/RoleController.php
@@ -11,6 +11,7 @@ use Spatie\Permission\Models\Role;
 use Spatie\Permission\Models\Permission;
 use App\Http\Resources\RoleResource;
 use App\Http\Requests\StoreRoleRequest;
+use App\Http\Requests\UpdateRoleRequest;
 use App\Http\Resources\PermissionResource;
 
 class RoleController extends Controller
@@ -58,7 +59,7 @@ class RoleController extends Controller
         try {
             DB::transaction(function () use ($validated) {
                 $role = Role::create(['name' => $validated['name']]);
-
+                $role->is_system = false;
                 $permissionIds = $validated['permissions'] ?? [];
                 $role->syncPermissions($permissionIds);
             });
@@ -92,9 +93,11 @@ class RoleController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(\Illuminate\Http\Request $request, string $id)
+    public function update(UpdateRoleRequest $request, string $id)
     {
         //
+        $validated = $request->validated();
+
     }
 
     /**
@@ -113,10 +116,13 @@ class RoleController extends Controller
 
                 // Se limpian relaciones antes de eliminar para mantener consistencia en pivotes.
                 $role->syncPermissions([]);
-                $role->delete();
+                if($role->is_system){
+                    return redirect()->back()->with('error', 'Este rol no se puede eliminar ya que es parte del sistema');
+                }else{
+                    $role->delete();
+                    return redirect()->back()->with('success', 'Rol eliminado correctamente.');
+                }
             });
-
-            return redirect()->back()->with('success', 'Rol eliminado correctamente.');
         } catch (ModelNotFoundException $e) {
             return redirect()->back()->withErrors(['role' => 'El rol que intentas eliminar no existe.']);
         } catch (\Throwable $e) {
